% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpCOVID.R
\name{CpCOVID}
\alias{CpCOVID}
\title{Basic description}
\usage{
CpCOVID(start,end,data)
}
\arguments{
\item{start, end}{starting point, ending point}

\item{data}{input data}
}
\value{
the change points estimation
}
\description{
Given the starting point, ending point, and the shortest Hamiltonian path, this function returns the change points estimation introduced in Shi, Chen, Dong and Rao (2020)
}
\examples{

# The data example of change points estimation in Shi, Chen, Dong and Rao (2020).
data(origin_data,package="GraphCpClust")
CpCOVID(1,142,Data.new[,"HB"])
}
\seealso{

}
