% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{Data.new}
\alias{Data.new}
\title{Data.new}
\description{
The log-transformed infectious counts form December 1,2019 to April 20,2020 applied in Shi, Chen, Dong and Rao (2020) for China's provinces/regions.
}
