% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{Data.new_country}
\alias{Data.new_country}
\title{Data.new_country}
\description{
The log-transformed infectious counts from December 1,2019 to April 20,2020 applied in Shi, Chen, Dong and Rao (2020) for 33 selected countries in the rest of world.
}
