% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance.data_cell_bee.R
\name{Distance.data_cell_bee}
\alias{Distance.data_cell_bee}
\title{Basic description}
\usage{
Distance.data_cell_bee(Data1,n1,n2,fun.dist)
}
\arguments{
\item{Data1}{input data}

\item{n1}{starting point/node}

\item{n2}{ending point/node}

\item{fun.dist}{weight function (Euclidean, lnorm, or may be defined by the reader)}
}
\value{
distance matrix - depends upon weight function what kind of distance it returns
}
\description{
Given the starting point, ending point and weight function, and this function returns the distance matrix
}
\examples{
# See Hpath_cell_bee example.
}
\seealso{
Hpath_cell_bee
}
