% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{final_C}
\alias{final_C}
\title{final_C}
\description{
The change points eatimation of each cluster of 33 selected countries in the rest of world introduced in Shi, Chen, Dong and Rao (2020).
}
