% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hpath_covid.R
\name{Hpath_covid}
\alias{Hpath_covid}
\title{Basic description}
\usage{
Hpath_covid(n1,n2,mat)
}
\arguments{
\item{n1}{starting node}

\item{n2}{ending node}

\item{mat}{distance matrix (distance type is determined by the reader)}
}
\value{
list of nodes of each edge from the shortest Hamiltonian path
}
\description{
Applies the path.kruskal function based on the nodes and edge.cost (sorts the weights from minimum to maximum). Given the starting node, ending node, and the distance matrix, this function returns the list of nodes of each edge from the shortest Hamiltonian path. We have the Hamiltonian path from path.kruskal
}
\examples{
data(origin_data,package="GraphCpClust")
Lnorm = function(x) sqrt(sum(t(x)*x))
data.trans=Distance.data_covid(t(Data.new),1,34,Lnorm)
HD=data.trans$hd # distance matrix
Hpath_covid(1,34,HD)
}
\seealso{
Distance.data_covid, path.kruskal_covid
}
