% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogPLotChina.R
\name{LogPLotChina}
\alias{LogPLotChina}
\title{Basic description}
\usage{
LogPLotChina(data_clust, final)
}
\arguments{
\item{data_clust}{data_clust}

\item{final}{final}
}
\description{
This function is used to fit the log-transformed infection counts of provinces/regions in China that introduced in Shi, Chen, Dong and Rao (2020)
}
\examples{

data("cluster_data",package = "GraphCpClust")

LogPLotChina(data_clust, final)
}
\seealso{

}
