% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogPLotCountry.R
\name{LogPLotCountry}
\alias{LogPLotCountry}
\title{Basic description}
\usage{
LogPLotCountry(data_clust_C, final_C)
}
\arguments{
\item{data_clust_C}{data_clust_C}

\item{final_C}{final_C}
}
\description{
This function is used to fit the log-transformed infection counts of 33 countries in the rest of world that introduced in Shi, Chen, Dong and Rao (2020)
}
\examples{

data("cluster_data",package="GraphCpClust")
LogPLotCountry(data_clust_C, final_C)
}
\seealso{

}
