% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.kruskal_covid.R
\name{path.kruskal_covid}
\alias{path.kruskal_covid}
\title{Basic description}
\usage{
path.kruskal_covid(nodes,edge_cost)
}
\arguments{
\item{nodes}{sequence of nodes 1,...,n from the graph which is based on the high-dimensional data that is provided by the reader}

\item{edge_cost}{sorted edge weights}
}
\value{
the shortest Hamiltonian path
}
\description{
Calculates the shortest Hamiltonian path based on the sorted edge weights and the nodes
}
\examples{
# path.kruskal is used in the Hpath function as follows:
Hpath_covid=function(n1,n2,mat){
n0=n2-n1+1
edge.cost=matrix(NA,nrow=n0*(n0-1)/2,ncol=3)

temp=1;

for(i in n1:(n2-1))
for(j in (i+1):(n2))
{
edge.cost[temp,3]=mat[i,j];edge.cost[temp,1]=i-n1+1;edge.cost[temp,2]=j-n1+1;temp=temp+1;}

edge.cost=edge.cost[sort.list(edge.cost[,3]), ]

#based on path

opt.path=path.kruskal_covid(c(1:n0),edge.cost)

return(opt.path$edge.cost[,1:2])
}
}
\seealso{
Hpath_covid
}
