% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationData.R
\name{SimulationData}
\alias{SimulationData}
\title{Basic description}
\usage{
SimulationData()
}
\description{
This function is used to generate simulation data introduced in Shi, Chen, Dong and Rao (2020)
}
\seealso{
Hpath_covid, Distance.data_covid, path.kruskal_covid
}
