% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationPlot.R
\name{SimulationPlot}
\alias{SimulationPlot}
\title{Basic description}
\usage{
SimulationPlot()
}
\description{
This fuction is the comparisons of graph-based clustering method and model-based clustering method introduced in Shi, Chen, Dong and Rao (2020)
}
\examples{
SimulationPlot()
}
\seealso{

}
